phastpp3_F90 should be equivalent to phastpp3_worker (9018) except for Fortran wrapper RM_interface.F90 is added.



git-svn-id: svn://136.177.114.72/svn_GW/phastpp/branches/phastpp3-F90@9019 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phast/PhreeqcRM/RM_interface_C.cpp
@@ -287,7 +287,7 @@ RM_GetDensity(int id, double * d)
 }
 /* ---------------------------------------------------------------------- */
 IRM_RESULT 
-RM_GetErrorString(int id, char *errstr, size_t l)
+RM_GetErrorString(int id, char *errstr, int l)
 /* ---------------------------------------------------------------------- */
 {
 	// Retrieves file prefix in prefix
